CI fix

+++ native/cocos/2d/renderer/RenderEntity.h
@@ -42,7 +42,7 @@ enum class RenderEntityType {
 };
 
 struct EntityAttrLayout {
-    float localOpacity{1.0f};
+    float localOpacity{1.0F};
     uint8_t colorR{255};
     uint8_t colorG{255};
     uint8_t colorB{255};
@@ -104,7 +104,7 @@ public:
     inline bool getColorDirty() const { return _entityAttrLayout.colorDirtyBit != 0; }
     inline void setColorDirty(bool dirty) { _entityAttrLayout.colorDirtyBit = dirty ? 1 : 0; }
     inline Color getColor() const { return Color(_entityAttrLayout.colorR, _entityAttrLayout.colorG, _entityAttrLayout.colorB, _entityAttrLayout.colorA); }
-    inline float getColorAlpha() const { return _entityAttrLayout.colorA / 255.F; }
+    inline float getColorAlpha() const { return static_cast<float>(_entityAttrLayout.colorA / 255.F); }
     inline float getLocalOpacity() const { return _entityAttrLayout.localOpacity; }
     inline float getOpacity() const { return _opacity; }
     inline void setOpacity(float opacity) { _opacity = opacity; }
@@ -132,7 +132,7 @@ private:
 
     bindings::NativeMemorySharedToScriptActor _entitySharedBufferActor;
 
-    float _opacity{1.0f};
+    float _opacity{1.0F};
 
     bool _isMask{false};
     bool _isSubMask{false};
